refactor: refactored notifications app and removed unused code
