+++ openedx/core/djangoapps/notifications/urls.py
@@ -11,7 +11,6 @@ from .views import (
     NotificationCountView,
     NotificationListAPIView,
     NotificationReadAPIView,
-    UserNotificationChannelPreferenceView,
     UserNotificationPreferenceView,
     preference_update_from_encrypted_username_view,
 )
@@ -25,11 +24,6 @@ urlpatterns = [
         fr'^configurations/{settings.COURSE_KEY_PATTERN}$',
         UserNotificationPreferenceView.as_view(),
         name='notification-preferences'
-    ),
-    re_path(
-        fr'^channel/configurations/{settings.COURSE_KEY_PATTERN}$',
-        UserNotificationChannelPreferenceView.as_view(),
-        name='notification-channel-preferences'
     ),
     path('', NotificationListAPIView.as_view(), name='notifications-list'),
     path('count/', NotificationCountView.as_view(), name='notifications-count'),
