+++ openedx/core/djangoapps/notifications/views.py
@@ -38,7 +38,7 @@ from .serializers import (
     NotificationCourseEnrollmentSerializer,
     NotificationSerializer,
     UserCourseNotificationPreferenceSerializer,
-    UserNotificationPreferenceUpdateSerializer, UserNotificationChannelPreferenceUpdateSerializer,
+    UserNotificationPreferenceUpdateSerializer,
 )
 from .utils import get_show_notifications_tray
 
@@ -230,55 +230,6 @@ class UserNotificationPreferenceView(APIView):
         updated_notification_preferences = preference_update.save()
         notification_preference_update_event(request.user, course_id, preference_update.validated_data)
 
-        serializer_context = {
-            'course_id': course_id,
-            'user': request.user
-        }
-        serializer = UserCourseNotificationPreferenceSerializer(updated_notification_preferences,
-                                                                context=serializer_context)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-
-
-@allow_any_authenticated_user()
-class UserNotificationChannelPreferenceView(APIView):
-    """
-    Supports retrieving and patching the UserNotificationPreference
-    model.
-    **Example Requests**
-        PATCH /api/notifications/configurations/{course_id}
-    """
-
-    def patch(self, request, course_key_string):
-        """
-        Update an existing user notification preference for an entire channel with the data in the request body.
-
-        Parameters:
-            request (Request): The request object
-            course_key_string (int): The ID of the course of the notification preference to be updated.
-        Returns:
-            200: The updated preference, serialized using the UserNotificationPreferenceSerializer
-            404: If the preference does not exist
-            403: If the user does not have permission to update the preference
-            400: Validation error
-        """
-        course_id = CourseKey.from_string(course_key_string)
-        user_course_notification_preference = CourseNotificationPreference.objects.get(
-            user=request.user,
-            course_id=course_id,
-            is_active=True,
-        )
-        if user_course_notification_preference.config_version != get_course_notification_preference_config_version():
-            return Response(
-                {'error': _('The notification preference config version is not up to date.')},
-                status=status.HTTP_409_CONFLICT,
-            )
-
-        preference_update = UserNotificationChannelPreferenceUpdateSerializer(
-            user_course_notification_preference, data=request.data, partial=True
-        )
-        preference_update.is_valid(raise_exception=True)
-        updated_notification_preferences = preference_update.save()
-        notification_preference_update_event(request.user, course_id, preference_update.validated_data)
         serializer_context = {
             'course_id': course_id,
             'user': request.user
